+++ module/chat-message/chat-message.js
@@ -7,21 +7,27 @@ import { actorSaveAction } from "../actions/actor-save-action.js";
  */
 export class MBChatMessage extends ChatMessage {
   /** @override */
-  async getHTML() {
-    const html = await super.getHTML();
+  async renderHTML() {
+    const html = await super.renderHTML();
     if (this.flags.systemMessage) {
       if (this.flags.cssClasses.value) {
-        html.addClass(this.flags.cssClasses.value);
+        html.classList.add(...this.flags.cssClasses.value);
       }
 
       const actor = ChatMessage.getSpeakerActor(this.speaker);
 
       if (actor?.sheet?.isEditable) {
-        html.on("click", "a.inline-roll", (event) => this.#onInlineRollClick(event));
-        html.on("click", "button.chat-message-button", (event) => this.#onButtonClick(event));
+        html.querySelectorAll("a.inline-roll").forEach(el => {
+          el.addEventListener("click", event => this.#onInlineRollClick(event));
+        });
+
+        html.querySelectorAll("button.chat-message-button").forEach(el => {
+          el.addEventListener("click", event => this.#onButtonClick(event));
+        });
+
       } else {
-        html.find("a.inline-roll").each((index, element) => $(element).attr("disabled", true));
-        html.find("button.chat-message-button").each((index, element) => $(element).attr("disabled", true));
+        html.querySelectorAll("a.inline-roll").forEach(el => el.setAttribute("disabled", "true"));
+        html.querySelectorAll("button.chat-message-button").forEach(el => el.setAttribute("disabled", "true"));
       }
     }
     return html;
@@ -32,6 +38,7 @@ export class MBChatMessage extends ChatMessage {
     event.stopPropagation();
 
     const actor = ChatMessage.getSpeakerActor(this.speaker);
+
     return (actor?.sheet?.isEditable) ? actorInlineRollAction(actor, this.#getOnlineRollData(event)) : null;
   }
 
@@ -39,11 +46,12 @@ export class MBChatMessage extends ChatMessage {
     event.preventDefault();
 
     const actor = ChatMessage.getSpeakerActor(this.speaker);
+
     return (actor?.sheet?.isEditable) ? this.#handleButtons(actor, event.currentTarget) : null;
   }
 
   #getEventData(event, data) {
-    return $(event.target).closest(`[data-${data}]`).data(data);
+    return event.target.closest(`[data-${data}]`)?.dataset[data];
   }
 
   /**