+++ module/models/actor/creature-data.js
@@ -1,0 +1,7 @@
+import { CommonData } from "./common-data.js";
+
+export class CreatureData extends CommonData {
+  static defineSchema() {
+    return super.defineSchema();
+  }
+}