+++ module/models/actor/warband-data.js
@@ -1,0 +1,7 @@
+import { CommonData } from "./common-data.js";
+
+export class WarbandData extends CommonData {
+  static defineSchema() {
+    return super.defineSchema({});
+  }
+}